Accept fractional Codex rate-limit percentages

+++ crates/model-runtime-codex-cli/src/app_server/frame.rs
@@ -286,7 +286,7 @@ pub(crate) struct RateLimits {
 #[derive(Clone, Debug, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub(crate) struct RateLimitWindow {
-    pub(crate) used_percent: i32,
+    pub(crate) used_percent: f64,
     pub(crate) resets_at: Option<i64>,
 }
 
@@ -304,7 +304,7 @@ impl RateLimits {
         let latest = [self.primary.as_ref(), self.secondary.as_ref()]
             .into_iter()
             .flatten()
-            .filter(|window| window.used_percent >= 100)
+            .filter(|window| window.used_percent >= 100.0)
             .filter_map(|window| window.resets_at)
             .max()?;
         let reset = std::time::UNIX_EPOCH.checked_add(std::time::Duration::from_secs(
